requirements.txt
aiohttp==3.11.14
discord.py-self==2.0.1